CFGEq.CFG

Description

The DSL used to describe CFGs.

Documentation

type CFG v t = ([v], [t], [Rule v t], t)

CFG v t is the type of context-free grammars with variables of type v and terminals of type t. Each such grammar is a 4-tuple of the list of variables, list of terminals, list of Rules and the start symbol.

data Rule v t

Rule v t is a CFG rule with variables of type v and terminals of type t. Variables with more than one rule should have multiple Rules, for example, \(S \to AB \mid \epsilon\) should have rules [S :-> [Left A, Left B], S :-> []].

Constructors

v :-> [Either v t]
S :-> s is the rule generating the string given by the list s from the variable S. The string is a list of Eithers, with Lefts representing variables and Rights representing terminals.

(.|) :: NonEmpty (Rule v t) -> [Either v t] -> NonEmpty (Rule v t)

Acts similarly to \(\mid\) in normal CFGs, adding a new rule to a NonEmpty list of existing rules. For example, \(S \to AB \mid c\) would be written S .-> [Left A, Left B] .| [Right 'c'].

(.->) :: v -> [Either v t] -> NonEmpty (Rule v t)

Creates a single rule as a NonEmpty list, allowing for use with .| to write multiple rules with the same terminal.