CFGEq.CNF

Description

Compilation of CFGs from CFG to Chomsky normal form.

Documentation

data CNF v t

CNF v t is the type of Chomsky normal form grammars with variables of type v and terminals of type t.

Constructors

CNF

Fields

rulesCNF :: Map v (Set (CNFProduction v t))

Map of variables to their set of rules.

startCNF :: v

The start symbol.

producesEmpty :: Bool

Whether the start symbol produces epsilon.

type CNFProduction v t = Either (v, v) t

The type of productions in a rule for CNF. Does not allow epsilon productions since CNF captures that for the start symbol as a Bool.

compile

Arguments

:: forall v t. (Ord v, Ord t)

=> (Int -> v)
Fresh variable supplier.

-> CFG v t
Grammar to compile.

-> CNF v t

Compiles a CFG to CNF. Requires the caller to supply a function that supplies fresh variables of the correct type, that should be guaranteed to return a different variable when given a different Int argument.

enumerate

Arguments

:: forall v t. (Ord v, Ord t)

=> Int
Maximum length of string to generate.

-> CNF v t
Chomsky normal form grammar to use.

-> Set [t]

enumerate n c returns the set of all strings generated by c of length n or less.